requirements.txt
# Core scientific stack
pandas>=2.0,<3
numpy>=1.24,<2
matplotlib>=3.7,<4
seaborn>=0.13,<0.14

# Machine Learning
scikit-learn>=1.3,<2
xgboost>=1.7,<3
joblib>=1.3,<2

# API (from v1.3 onwards)
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<1

# Database (added in v1.5)
psycopg2-binary>=2.9,<3
sqlalchemy>=2.0,<3
alembic>=1.13,<2

# Env loader
python-dotenv>=1.0,<2
